feat (items): add file type validation and slight refactoring - created a subfolder for file upload options and utils - extracted callback from storage-options to file.utils
